Remove node/src/proposal.rs
- Inline create_macro_block_proposal() into propose_macro_block().
- Move validate_proposed_macro_block() to blockchain/src/validation.rs
- Merge NodeBlockError into BlockError.

This patch is pure refactoring. No semantic changes.
