+++ blockchain/src/error.rs
@@ -359,6 +359,21 @@ pub enum BlockError {
         _0, _1, _2, _3
     )]
     OutOfSyncTimestamp(u64, Hash, Timestamp, Timestamp),
+    #[fail(
+        display = "Invalid block proposal: epoch={}, expected={}, got={}",
+        _0, _1, _2
+    )]
+    InvalidBlockProposal(u64, Hash, Hash),
+    #[fail(
+        display = "Invalid block epoch found: block_epoch={}, chain_epoch={}",
+        _0, _1
+    )]
+    InvalidBlockEpoch(u64, u64),
+    #[fail(
+        display = "Proposed view_change different from ours: epoch={}, block={}, block_viewchange={}, our_viewchange={}",
+        _0, _1, _2, _3
+    )]
+    OutOfSyncViewChange(u64, Hash, u32, u32),
 }
 
 #[derive(Debug, Fail)]
