+++ node/src/error.rs
@@ -52,25 +52,6 @@ pub enum NodeTransactionError {
     MempoolIsFull(Hash),
 }
 
-#[derive(Debug, Fail, PartialEq, Eq)]
-pub enum NodeBlockError {
-    #[fail(
-        display = "Invalid block proposal: epoch={}, expected={}, got={}",
-        _0, _1, _2
-    )]
-    InvalidBlockProposal(u64, Hash, Hash),
-    #[fail(
-        display = "Invalid block epoch found: block_epoch={}, chain_epoch={}",
-        _0, _1
-    )]
-    InvalidBlockEpoch(u64, u64),
-    #[fail(
-        display = "Proposed view_change different from ours: epoch={}, block={}, block_viewchange={}, our_viewchange={}",
-        _0, _1, _2, _3
-    )]
-    OutOfSyncViewChange(u64, Hash, u32, u32),
-}
-
 #[derive(Debug, Fail)]
 pub enum ForkError {
     #[fail(display = "Our branch is more significant, drop this block.")]
@@ -82,12 +63,6 @@ pub enum ForkError {
 impl From<failure::Error> for ForkError {
     fn from(err: failure::Error) -> ForkError {
         ForkError::Error(err)
-    }
-}
-
-impl From<NodeBlockError> for ForkError {
-    fn from(err: NodeBlockError) -> ForkError {
-        ForkError::Error(err.into())
     }
 }
